Clear the word list cache
Word lists for items are loaded and cached separatly from the item they
are associated with. The Blackbox WordListPanel does not have a way to
clear the cached word list. I added a method to the WordListPanel to clear
the cache so items can be reloaded with a different word list. This is
used for the glossary accommodation.