Dictionary scripts and button
The dictionary button is only enabled when the dictionary accommodation is
loaded. The dictionary accommodation is not yet fully implemented. It
makes an API call to a dictionary URL that is defined in the settings.
Where that is set is not clear so the API call will be implemented at a
later date.