Moved accom JS to a blackbox resource file
Some of the accommodations required writing or extending the JavaScript
from Blackbox. They were being loaded in Iris in an attempt to keep
modifications out of the blackbox. The scripts have been moved into
Blackbox with the scripts from student in order to follow the way the
Student module handles its scripts.