Updated POM with newer dependencies and exclusions
The goal is to not require changes to the item renderer to be compiled into
student before being compiled into Iris. We still need to depend on student for
some things, but we are now pulling the the item-renderer and item-renderer-web
separatley, and no longer depending on the student.library.